WIP: Big Decimal support for Numeric types in postgres

+++ src/models.rs
@@ -1,11 +1,14 @@
+use super::schema::accession_numbers;
+use bigdecimal::BigDecimal;
+
 #[derive(Queryable)]
 pub struct AccessionNumber {
     pub id: i32,
-    pub accession_number: u64,
+    pub accession_number: BigDecimal,
 }
 
 #[derive(Insertable)]
 #[table_name = "accession_numbers"]
 pub struct NewAccessionNumber {
-    pub accession_number: u64,
+    pub accession_number: BigDecimal,
 }
